\name{mni.split.data.for.nodes}
\alias{mni.split.data.for.nodes}
\title{Sets up data for parallel processing of dataset}
\description{
  Splits the data into smaller chunks that can then be processed separately.
}
\usage{
mni.split.data.for.nodes(data.table, node.setup, directory)
}
\arguments{
  \item{data.table}{The data table generated by \link{mni.build.data.table}.}
  \item{node.setup}{The processing instructions set up by
    \link{mni.rows.per.node}.}
  \item{directory}{The directory into which the data segments will be written.}
}
\value{
  Returns an array of filenames, one entry per node, giving the full
  path to the data for that node.
}
\seealso{
  \code{\link{mni.build.data.table}}.
  \code{\link{mni.rows.per.node}}.
}
