\name{mni.vertex.anova}
\alias{mni.vertex.anova}
\title{Per vertex ANOVA}
\description{
  An ANOVA at each vertex.
}
\usage{
va <- mni.vertex.anova(glim.matrix, statistics.model, vertex.table=FALSE)
}
\arguments{
  \item{glim.matrix}{The design matrix as specified by
    \link{mni.read.glim.file}}
  \item{statistics.model}{The formula to test}
  \item{vertex.table}{The data table as generated by
    \link{mni.build.data.table}}
}
\value{
  Returns a matrix with the F values for each element in the formula
}
\details{
  Runs an ANOVA at each vertex. See \link{aov} for more details about
  what to expect.
}
\seealso{
  \link{mni.vertex.statistics}
  \link{aov}
}
\examples{
}
\author{}
\keyword{regression}